formatting of decimals

+++ src/wrapped_radix_engine_toolkit/high_level/sargon_specific_types/assets_transfers/per_asset/per_asset_fungible_transfer.rs
@@ -15,7 +15,7 @@ impl PerAssetFungibleTransfer {
 
     pub fn amount(
         &self,
-        divisibility: impl Into<Option<i32>>,
+        divisibility: impl Into<Option<u8>>,
     ) -> ScryptoDecimal192 {
         self.amount.round(divisibility).into()
     }
